Remove keypress check and use render on/off count
NOTE: If multi-threading is on then Box2 must be OFF i.e.
not drawn, for more than 1 frame e.g. a flip/flop
of render/don't render box2 will work with multi-threading,
but crash with single threaded bgfx.

If Box2 is OFF (not drawn) for more than 2 frames then the
crash happens regardless of threading.

ON is assumed to be > 0 in all of the above i.e. Box2 must
be drawn at some point.